+++ Domain/Contracts/OrderItem.cs
@@ -13,9 +13,9 @@ namespace Domain.Contracts
     {
         public int Quantity { get; set; }
 
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
-        public decimal TotalPrice { get; set; }
+        public int TotalPrice { get; set; }
 
         public string OrderId { get; set; }
